+++ src/routes/Authorize/AuthorizeOtp.tsx
@@ -1,5 +1,5 @@
 import { observer } from 'mobx-react-lite';
-import { useCallback } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { Stack, useIsMobile, useTheme, ArrowBackIosIcon } from '@cere-wallet/ui';
 
 import { useLocation, useNavigate, useOutletContext } from 'react-router-dom';
@@ -8,18 +8,28 @@ import { OtpPage } from '~/components';
 import { AuthorizePopupStore } from '~/stores';
 import { useAppContextStore } from '~/hooks';
 
-const AuthorizeOtp = () => {
+type AuthorizeOtpProps = {
+  sendOtp?: boolean;
+};
+
+const AuthorizeOtp = ({ sendOtp }: AuthorizeOtpProps) => {
   const isMobile = useIsMobile();
   const location = useLocation();
   const navigate = useNavigate();
+  const [isBusy, setBusy] = useState(false);
+  const [autoOtp, setAutoOtp] = useState<string>();
   const store = useOutletContext<AuthorizePopupStore>();
   const { whiteLabel } = useAppContextStore();
   const { isGame } = useTheme();
+  const hasBackButton = !isGame && !!location.state?.email;
 
-  store.email = location.state?.email;
+  if (location.state?.email) {
+    store.email = location.state?.email;
+  }
 
   const handleLoginRequest = useCallback(
     async (idToken: string) => {
+      setBusy(true);
       const { isNewUser } = await store.login(idToken);
 
       if (
@@ -35,9 +45,22 @@ const AuthorizeOtp = () => {
       }
 
       await store.acceptSession();
+      setBusy(false);
     },
     [location, navigate, store, whiteLabel],
   );
+
+  useEffect(() => {
+    if (sendOtp) {
+      store.sendOtp();
+    }
+
+    store.waitForAuthLinkToken(async ({ token, code }) => {
+      setAutoOtp(code);
+
+      await handleLoginRequest(token);
+    });
+  }, [handleLoginRequest, sendOtp, store]);
 
   if (isMobile) {
     return (
@@ -49,9 +72,10 @@ const AuthorizeOtp = () => {
         height="100vh"
         spacing={9}
       >
-        {isGame ? null : <ArrowBackIosIcon onClick={() => navigate(-1)} />}
+        {hasBackButton && <ArrowBackIosIcon onClick={() => navigate(-1)} />}
+
         <Stack direction="column" textAlign="justify">
-          <OtpPage email={store.email} onRequestLogin={handleLoginRequest} />
+          <OtpPage code={autoOtp} busy={isBusy} email={store.email} onRequestLogin={handleLoginRequest} />
         </Stack>
       </Stack>
     );
@@ -65,10 +89,11 @@ const AuthorizeOtp = () => {
       padding={2}
       height="100vh"
     >
-      {isGame ? null : <ArrowBackIosIcon onClick={() => navigate(-1)} />}
+      {hasBackButton && <ArrowBackIosIcon onClick={() => navigate(-1)} />}
+
       <Stack direction="row" justifyContent="center" alignItems="center" padding={2} height="100vh">
         <Stack width={375}>
-          <OtpPage email={store.email} onRequestLogin={handleLoginRequest} />
+          <OtpPage code={autoOtp} busy={isBusy} email={store.email} onRequestLogin={handleLoginRequest} />
         </Stack>
       </Stack>
     </Stack>